pyproject.toml
[project]
name = "ai-search-agent-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "asyncpg>=0.30.0",
    "google-api-python-client>=2.187.0",
    "openai>=2.7.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
